* origin/master:
  doc/import-rsnapshot: small corrections and clarification
  cmd/midx, git.py: all else being equal, delete older midxes first.
  t/test.sh: a test for the recently-uncovered midx4 problem.
  _helpers.c: midx4 didn't handle multiple index with the same object.
  cmd/midx: add a --check option.
  Add git.shorten_hash(), printing only the first few bytes of a sha1.
  tclient.py: add some additional tests that objcache.refresh() is called.
  cmd/server: add a debug message saying which object caused a suggestion.
  cmd/list-idx: a quick tool for searching the contents of idx/midx files.
  Add tests around the bloom ruin and check options
  Add a bloom --ruin for testing failure cases
  One more constant for header lengths
  Split PackMidx from git.py into a new midx.py.
  bloom.py: move bloom.ShaBloom.create to just bloom.create.
  Move bloom-related stuff from git.py to a new bloom.py.
  cmd/bloom: add a --force option to forget regenerating the bloom.
  Use the new qprogress() function in more places.
  Bail out immediately instead of redownloading .idx
  Add a --check behavior to verify bloom
  Defines/preprocessor lengths > magic numbers
  cmd/{bloom,midx}: clean up progress messages.
  cmd/bloom: by default generate bloom filters in *all* index-cache dirs.
  cmd/newliner: avoid printing a blank line if the final output ended in \r.
  cmd/index: make the default mode '-u'.
  _helpers.c: don't cygwin doesn't set any win32 defines.
  _helpers.c: don'g unpythonize_argv() on win32.
  Remove .c and .o rules, apply flags to csetup.py
  Fix a valid warning that some compilers warned
  Move .idx file writing to C
  main.py: fix whitespace in the usage string.
  cmd/daemon: FD_CLOEXEC the listener socket and don't leak fd for the connection.
  cmd/daemon: close file descriptors correctly in parent process.
  cmd/daemon: use SO_REUSEADDR.
  cmd/daemon: pass extra options along to 'bup server'.
  cmd/daemon: correctly report socket binding/listening errors.
  main.py: use execvp() instead of subprocess.Popen() when possible.
  _helpers.c: Remove ugly 'python' junk from argv[0] so 'ps' is prettier.
  cmd/bloom: fix a message pluralization.
  cmd/join: add a new -o (output filename) option.
  cmd/ls: fix a typo causing 'bup ls foo/latest' to not work.
  cmd/server: add a new 'help' command.
  midx4: Fix the other side of my previous nasty bug

Conflicts:
	lib/bup/_helpers.c
	lib/bup/helpers.py

+++ lib/bup/client.py
@@ -182,17 +182,20 @@ class Client:
         #debug1('requesting %r\n' % name)
         self.check_busy()
         mkdirp(self.cachedir)
+        fn = os.path.join(self.cachedir, name)
+        if os.path.exists(fn):
+            msg = "won't request existing .idx, try `bup bloom --check %s`" % fn
+            raise ClientError(msg)
         self.conn.write('send-index %s\n' % name)
         n = struct.unpack('!I', self.conn.read(4))[0]
         assert(n)
-        fn = os.path.join(self.cachedir, name)
         f = open(fn + '.tmp', 'w')
         count = 0
         progress('Receiving index from server: %d/%d\r' % (count, n))
         for b in chunkyreader(self.conn, n):
             f.write(b)
             count += len(b)
-            progress('Receiving index from server: %d/%d\r' % (count, n))
+            qprogress('Receiving index from server: %d/%d\r' % (count, n))
         progress('Receiving index from server: %d/%d, done.\n' % (count, n))
         self.check_ok()
         f.close()
@@ -213,11 +216,13 @@ class Client:
             debug2('%s\n' % line)
             if line.startswith('index '):
                 idx = line[6:]
-                debug1('client: received index suggestion: %s\n' % idx)
+                debug1('client: received index suggestion: %s\n'
+                       % git.shorten_hash(idx))
                 suggested.append(idx)
             else:
                 assert(line.endswith('.idx'))
-                debug1('client: completed writing pack, idx: %s\n' % line)
+                debug1('client: completed writing pack, idx: %s\n'
+                       % git.shorten_hash(line))
                 suggested.append(line)
         self.check_ok()
         if ob:
